Version1-Shivani/requirements.txt
Flask
indic-transliteration
pytesseract
Pillow
